+++ legacy/src/main/java/org/openscience/cdk/aromaticity/CDKHueckelAromaticityDetector.java
@@ -180,7 +180,7 @@ public class CDKHueckelAromaticityDetector {
                     IAtomType type = factory.getAtomType(ringAtom.getAtomTypeName());
                     Object property = type.getProperty(CDKConstants.PI_BOND_COUNT);
                     if (property != null && property instanceof Integer) {
-                        electronCount += ((Integer) property).intValue();
+                        electronCount += (Integer) property;
                     }
                 }
             } else if (ringAtom.getHybridization() != null && ringAtom.getHybridization() == Hybridization.SP3
