+++ legacy/src/main/java/org/openscience/cdk/smsd/algorithm/matchers/DefaultVFBondMatcher.java
@@ -155,7 +155,7 @@ public class DefaultVFBondMatcher implements VFBondMatcher {
     }
 
     private int getValency(IAtom atom) {
-        return (atom.getValency() == null) ? 0 : atom.getValency().intValue();
+        return (atom.getValency() == null) ? 0 : atom.getValency();
     }
 
     private int getUnsaturation(IAtomContainer container, IBond bond) {
