+++ legacy/src/main/java/org/openscience/cdk/smsd/algorithm/rgraph/CDKMCSHandler.java
@@ -180,8 +180,8 @@ public class CDKMCSHandler extends AbstractMCSAlgorithm implements IMCSBase {
                 TreeMap<Integer, Integer> atomMappings = new TreeMap<>();
                 for (Map.Entry<Integer, Integer> solutions : finalSolution.entrySet()) {
 
-                    int iIndex = solutions.getKey().intValue();
-                    int jIndex = solutions.getValue().intValue();
+                    int iIndex = solutions.getKey();
+                    int jIndex = solutions.getValue();
 
                     if (rOnPFlag) {
                         atomMappings.put(iIndex, jIndex);
@@ -208,8 +208,8 @@ public class CDKMCSHandler extends AbstractMCSAlgorithm implements IMCSBase {
             Map<IAtom, IAtom> atomMappings = new HashMap<>();
             for (Map.Entry<Integer, Integer> solutions : finalSolution.entrySet()) {
 
-                int iIndex = solutions.getKey().intValue();
-                int jIndex = solutions.getValue().intValue();
+                int iIndex = solutions.getKey();
+                int jIndex = solutions.getValue();
 
                 IAtom sourceAtom = null;
                 IAtom targetAtom = null;
