+++ storage/io/src/test/java/org/openscience/cdk/io/cml/CML23FragmentsTest.java
@@ -129,7 +129,7 @@ public class CML23FragmentsTest extends CDKTestCase {
         Assert.assertEquals(1, mol.getAtomCount());
         IAtom atom = mol.getAtom(0);
         Assert.assertEquals("C", atom.getSymbol());
-        Assert.assertEquals(12.0, atom.getExactMass().doubleValue(), 0.01);
+        Assert.assertEquals(12.0, atom.getExactMass(), 0.01);
     }
 
     @Test
